requirements.txt
paho-mqtt
requests==2.28.1
mysql-connector-python
